# src/server/src/resources/swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "This is the API contract for the server for the DMS LEAP application",
    "version": "1.0.0",
    "title": "LEAP DMS APIs",
    "termsOfService": ""
  },
...
  "paths": {
...
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Register new user",
        "description": "Register new user based on data passed in the request body",
        "operationId": "createUser",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Users"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/users/byEmailId/{emailId}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by email Id",
        "description": "Get details of user based on its email ID",
        "operationId": "getUserByEmailId",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "emailId",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ByEmailId"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/users/byEVC/{code}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by EVC",
        "description": "Get details of user based on its EVC",
        "operationId": "getUserByEVC",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "code",
            "in": "path",
            "description": "code of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ByEVC"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
    "/users/{id}": {
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get user by Id",
        "description": "Get details of user based on its ID",
        "operationId": "getUserById",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserById"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Updates user information",
        "description": "Updates the user information",
        "operationId": "updateUser",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "example": "user@test.com"
                  },
                  "invitationCode": {
                    "type": "string",
                    "example": "MAFB4326"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "updated successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/users/{id}/acceptConsent": {
      "put": {
        "tags": [
          "Users"
        ],
        "summary": "Consent an app",
        "description": "adds ref of app associated with appId to user with isActive true",
        "operationId": "acceptConsent",
        "parameters": [
          {
            "in": "header",
            "description": "userId",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "ID of user to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "in": "query",
            "description": "application name (Study App)",
            "name": "source",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "platform",
            "required": true,
            "description": "type of platform (web, mobile)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "appId": {
                    "type": "string",
                    "example": "5ff53d85af962933a8b3a555"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcceptConsent"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          },
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "https://leapapi-dev.leap.dellmed.utexas.edu/api"
    },
    {
      "url": "https://localhost:8080/api"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "description": "user details",
        "properties": {
          "firstName": {
            "type": "string",
            "description": "First name of user",
            "example": "firstname"
          },
          "lastName": {
            "type": "string",
            "description": "Last name of user",
            "example": "lastname"
          },
          "birthday": {
            "type": "string",
            "description": "Date of birth of user",
            "example": "10/05/1992"
          },
          "email": {
            "type": "string",
            "description": "Email ID of user",
            "example": "test@test.com"
          },
          "gender": {
            "type": "string",
            "example": "male"
          },
          "phoneNumberPrimary": {
            "type": "number",
            "example": 1234567890
          },
          "phoneNumberSecondary": {
            "type": "number",
            "example": 1234567890
          },
          "invitationCode": {
            "type": "string",
            "example": "MAFB4326"
          },
          "zip": {
            "type": "number",
            "example": 123456
          },
          "role": {
            "type": "string",
            "example": "5ff40c7e153fc3d42838f9cb"
          },
          "createdBy": {
            "type": "string",
            "example": "user name"
          },
          "updatedBy": {
            "type": "string",
            "example": "user name"
          }
        }
      },
...
    },
    "responses": {
      "UnauthorizedError": {
        "description": "Access token is missing or invalid"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ]
}